pyproject.toml
[project]
name = "intelligent-synthesizer-ai"
version = "0.1.0"
description = "Agentic AI system for deep topic research and intelligent knowledge synthesis"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "gradio>=5.41.1",
    "ipywidgets>=8.1.7",
    "openai>=1.99.1",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.8.3",
    "pydantic>=2.10.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.30.1",
    "pytest>=8.4.1",
]
